feat(dvops): assembled DvOps into one file class
