+++ Derivas/Expression/Operators/Binary.cs
@@ -48,17 +48,4 @@ namespace Derivas.Expression
             }
         }
     }
-
-    public static partial class DvOps
-    {
-        public static IDvExpr Div(object fst, object snd)
-            => new BinaryOperator(CheckExpr(fst), CheckExpr(snd), "/", 1, (fst, snd) => fst / snd);
-
-        public static IDvExpr Sub(object fst, object snd)
-            => new BinaryOperator(CheckExpr(fst), CheckExpr(snd), "-", 0, (fst, snd) => fst - snd);
-
-        public static IDvExpr Pow(object bas, object pow) => new BinaryOperator(
-            CheckExpr(bas), CheckExpr(pow), "^", 2, (bas, pow) => Math.Pow(bas, pow)
-        );
-    }
 }
