+++ Derivas/Expression/Trivial.cs
@@ -47,10 +47,4 @@ namespace Derivas.Expression
         protected override CloneableExpr CreateFromClonable(params CloneableExpr[] expr)
             => expr.Length == 0 ? new Symbol(Name) : expr[0].CreateInstance();
     }
-
-    public static partial class DvOps
-    {
-        public static IDvExpr Const(double val) => new Constant(val);
-        public static IDvExpr Sym(string name) => new Symbol(name);
-    }
 }
