+++ DerivasTests/Expression/Operators/Binary.cs
@@ -1,10 +1,11 @@
 using Derivas.Exception;
 using NUnit.Framework;
 using System.Collections.Generic;
-using static Derivas.Expression.DvOps;
 
 namespace Derivas.Expression.Tests
 {
+    using static DvOps;
+
     [TestFixture]
     public class BinaryOpeartorTests : OperatorsTestUtility
     {
